sb_website/requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-countries==7.2.1
django-crispy-forms==1.13.0
django-import-export==2.7.1
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
pyuca==1.2
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
xlrd==2.0.1
xlwt==1.3.0
